Implement recursive upload and download of directories if they are specified in an `Application`'s `input` or `output` attribute.
Also, state as a policy that missing output files do not cause an
error (this is what ARC does, we propagate it to other backends as
well).

Other fixes:
  * gc3libs/__init__.py: 
    - Update `Application` docstring
    - New constant `ANY_OUTPUT` for specifying that we want the entire output sandbox back
    - Force `input` and `output` keys and values into ASCII strings; raise enconding/decoding exceptions if they are not.
  * gc3libs.backends.sge: only symlink `stderr_filename` to `stdout_filename` if join != True.
  * gc3libs.transport: 
    - Added `isdir` method.
    - `get` and `put` now behave recursively.


modified:
  gc3pie/gc3libs/__init__.py
  gc3pie/gc3libs/backends/arc.py
  gc3pie/gc3libs/backends/sge.py
  gc3pie/gc3libs/backends/subprocess.py
  gc3pie/gc3libs/backends/transport.py

+++ trunk/gc3pie/gc3libs/backends/sge.py
@@ -625,30 +625,29 @@ class SgeLrms(LRMS):
                     # ...but keep it if it's not a known one
                     remote_path = os.path.join(job.ssh_remote_folder, remote_path)
                 local_path = os.path.join(download_dir, local_path)
-                log.debug("Downloading remote file '%s' to local file '%s'"
-                          % remote_path, local_path)
+                log.debug("Downloading remote file '%s' to local file '%s'",
+                          remote_path, local_path)
                 try:
-                    if not os.path.exists(local_path) or overwrite:
+                    if (overwrite
+                        or not os.path.exists(local_path)
+                        or os.path.isdir(local_path)):
                         log.debug("Copying remote '%s' to local '%s'"
                                   % (remote_path, local_path))
-                        self.transport.get(remote_path, local_path)
+                        # effectively ignore missing files (this is
+                        # what ARC does too)
+                        self.transport.get(remote_path, local_path,
+                                           ignore_nonexisting=True)
                     else:
-                        log.info("Local file '%s' already exists; will not be overwritten!",
+                        log.info("Local file '%s' already exists;"
+                                 " will not be overwritten!",
                                  local_path)
-                except:
-                    log.error('Could not copy remote file: ' + remote_path)
-                    # FIXME: should we set `job.signal` to
-                    # `Job.Signals.DataStagingError`?  Does not seem a
-                    # good idea: What if one file is missing out of
-                    # several good ones?  Fetching output could be
-                    # attempted in any case, and (re)setting
-                    # `job.signal` here could mask the true error
-                    # cause.
+                except Exception:
                     raise
 
             # make jobname.stderr a link to jobname.stdout in case
             # some program relies on its existence
-            if not os.path.exists(os.path.join(download_dir, job.stderr_filename)):
+            if (not app.join
+                and not os.path.exists(os.path.join(download_dir, job.stderr_filename))):
                 os.symlink(os.path.join(download_dir, job.stdout_filename),
                            os.path.join(download_dir, job.stderr_filename))
